+++ sunset/bunch.py
@@ -357,8 +357,7 @@ class Bunch(ContainableImpl):
 
         self._update_notification_callbacks.callAll(field)
 
-        container = self._container()
-        if container is not None and not self.skipOnSave():
+        if (container := self._container()) is not None:
             container._triggerUpdateNotification(field)
 
     def _typeHint(self) -> type: